adjust SHAP algorithm to use precomputed number samples

+++ lib/maths/CBoostedTreeImpl.cc
@@ -472,7 +472,7 @@ CBoostedTreeImpl::trainForest(core::CDataFrame& frame,
     TNodeVecVec forest{this->initializePredictionsAndLossDerivatives(
         frame, trainingRowMask, testingRowMask)};
     forest.reserve(m_MaximumNumberTrees);
-    forest[0][0].numberSamples(trainingRowMask.manhattan());
+    forest[0][0].numberSamples(static_cast<std::size_t>(trainingRowMask.manhattan()));
 
     CScopeRecordMemoryUsage scopeMemoryUsage{forest, recordMemoryUsage};
